requirements.txt
Flask==2.3.3
tensorflow==2.15.0
keras==2.15.0
opencv-python==4.9.0.80
numpy==1.24.4
pillow==10.3.0
gunicorn==21.2.0